+++ src/Environment.php
@@ -27,6 +27,10 @@ use League\CommonMark\Util\ConfigurationAwareInterface;
 
 class Environment
 {
+    const HTML_INPUT_STRIP = 'strip';
+    const HTML_INPUT_ESCAPE = 'escape';
+    const HTML_INPUT_ALLOW = 'allow';
+
     /**
      * @var ExtensionInterface[]
      */
@@ -478,7 +482,8 @@ class Environment
                 'inner_separator' => "\n",
                 'soft_break'      => "\n",
             ],
-            'safe' => false,
+            'safe'       => false,
+            'html_input' => self::HTML_INPUT_ALLOW,
         ]);
 
         return $environment;